requirement.txt
alembic==1.12.0
BareNecessities==0.2.8
blinker==1.6.3
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
dnspython==2.4.2
dominate==2.8.0
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.0
idna==3.4
importlib-metadata==6.8.0
install==1.3.5
is-disposable-email==1.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mail==2.1.0
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.1.0
wtforms-validators==1.0.0
zipp==3.17.0
